[TEST] Merge Gluon tensor memory coverage
